// src/google/types/game-translation-type.ts
export enum GameTranslationType {
  ITEM = 'item',
  ITEM_DESC = 'item_desc',
  MOB = 'mob',
  SKILL = 'skill',
  LOCALE_STRING = 'locale_string',
  LOCALE_GAME = 'locale_game',
  LOCALE_CLIENT = 'locale_client',
  LOCALE_INTERFACE = 'locale_interface',
  ITEM_DESCRIPTION = 'item_description',
  EXTENDED_PET_SKILL = 'extended_pet_skill',
}